requirements.txt
groq>=0.4.0
sentence-transformers>=2.2.0
chromadb>=0.5.0
streamlit>=1.35.0
python-dotenv>=1.0.0
langchain-core>=0.2.0
langchain-text-splitters>=0.2.0
langchain-chroma>=0.1.0